Update copyright years and enhance documentation:
- Update copyright years across all source files to 2025.
- Enhance README.md to document multi-threaded support:
  - BZIP2: supports both multi-threaded compression and decompression
  - LZMA: supports multi-threaded compression only
- Adjust maximum destination buffer size for BZIP2 in codec_bench.h.
- Update Sphinx documentation to clarify multi-threading API support for BZIP2 and LZMA formats.

+++ test/codec_bench.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2022-2024, Advanced Micro Devices. All rights reserved.
+ * Copyright (C) 2022-2025, Advanced Micro Devices. All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -101,7 +101,7 @@ static const codec_list_t codec_list[AOCL_COMPRESSOR_ALGOS_NUM] =
     {"LZ4",    0, 0,  0, 0, 1, 0, 1, 1, 0, INT_MAX  , ".lz4"},
     {"LZ4HC",  1, 12, 0, 0, 1, 0, 1, 1, 0, INT_MAX  , ".lz4"},
     {"LZMA",   0, 9,  0, 0, 1, 0, 1, 0, 0, SIZE_MAX , ".lzma"},
-    {"BZIP2",  1, 9,  0, 0, 1, 0, 1, 1, 0, SIZE_MAX , ".bz2"},
+    {"BZIP2",  1, 9,  0, 0, 1, 0, 1, 1, 0, UINT_MAX , ".bz2"},
     {"SNAPPY", 0, 0,  0, 0, 1, 0, 1, 1, 0, SIZE_MAX , ".snappy"},
     {"ZLIB",   1, 9,  0, 0, 1, 0, 1, 1, 0, ULONG_MAX, ".zlib"},
     {"ZSTD",   1, 22, 0, 0, 1, 1, 1, 1, 1, SIZE_MAX , ".zst"}
